refactor: extract _atomic_write into shared src/tools/_io.py (#40)
